+++ NEWS.md
@@ -1,5 +1,6 @@
 # mlr3oml 0.7.0
 
+* fix: strings and nominals are distinguished for parquet files
 * docs: Fixed some OpenML links
 * docs: Renamed the docs for OpenML objects
 * Renamed the sugar functions from: